refactor: css 분리 #9
